Smoke test for JENKINS_PASSWORD
