+++ smoke/features/persistent-test.feature
@@ -10,5 +10,5 @@ Feature: Upon jenkins master pod deletion/destruction the data(jobs) persist whe
       Given The jenkins pod is up and runnning
       Then We rsh into the master pod and check the jobs count
       When We delete the jenkins master pod
-      Then We check for jenkins master pod status to be "Ready"
+      Then We ensure that jenkins deployment config is ready
       And We rsh into the master pod & Compare if the data persist or is lost upon pod restart
