+++ smoke/features/steps/debug.py
@@ -1,0 +1,9 @@
+from behave import given, then, when
+
+@when(u'true')
+def step_impl(context):
+    print( "fake when")
+
+@then(u'fail')
+def step_impl(context):
+    raise AssertionError
