+++ smoke/features/steps/steps.py
@@ -2,49 +2,41 @@
 # ----------------------------------------------------------------------------
 # STEPS:
 # ----------------------------------------------------------------------------
+import json
 import os
 import time
+
+import jenkins
+
 import urllib3
-from behave import given, when, then
+from behave import given, then, when
+from kubernetes import client, config
 from pyshould import should
-from kubernetes import config, client
 from smoke.features.steps.openshift import Openshift
-from smoke.features.steps.project import Project
 from smoke.features.steps.plugins import Plugins
+from smoke.features.steps.project import Project
 
 # Test results file path
 scripts_dir = os.getenv('OUTPUT_DIR')
 
-# Path to pipeline job to test agent images
-maven_template ='./smoke/samples/maven_pipeline.yaml'
-nodejs_template = './smoke/samples/nodejs_pipeline.yaml'
-
 # variables needed to get the resource status
 deploy_pod = "jenkins-1-deploy"
-jenkins_master_pod = ''
+global jenkins_master_pod
+global current_project
 current_project = ''
 config.load_kube_config()
-v1 = client.CoreV1Api()
 oc = Openshift()
 podStatus = {}
-buildconfigs = {'sample-pipeline':'1','openshift-jee-sample':'1'}
-builds = {}
 # Parse the base plugins from the file and store them in a dictonary with key=plugin-name & value=plugin-version
 
 baseplugins = './2/contrib/openshift/base-plugins.txt'
 p = Plugins()
 plugins = p.getPlugins(baseplugins)
 
-def triggerbuild(buildconfig,namespace):
-    print('Triggering build: ',buildconfig)
-    res = oc.start_build(buildconfig,namespace)
-    print(res)
-
 # STEP
 @given(u'Project "{project_name}" is used')
 def given_project_is_used(context, project_name):
     project = Project(project_name)
-    global current_project
     current_project = project_name
     context.current_project = current_project
     context.oc = oc
@@ -58,7 +50,7 @@ def given_project_is_used(context, project_name):
 
 def before_feature(context, feature):
     if scenario.name != None and "TEST_NAMESPACE" in scenario.name:
-        print("Scenario using env namespace subtitution found: {0}, env: {}".format(scenario.name, os.getenv("TEST_NAMESPACE"))
+        print("Scenario using env namespace subtitution found: {0}, env: {}".format(scenario.name, os.getenv("TEST_NAMESPACE")))
         scenario.name = txt.replace("TEST_NAMESPACE", os.getenv("TEST_NAMESPACE"))
 
 # STEP
@@ -69,20 +61,9 @@ def given_namespace_from_env_is_used(context, project_env):
     print(f"{project_env} = {env}")
     given_project_is_used(context, env)
 
-
-@given(u'we have a openshift cluster')
-def loginCluster(context):
-    print("Using [{}]".format(current_project))
-
-@when(u'User enters oc new-app jenkins-ephemeral command')
-def ephemeralTemplate(context):
-    res = oc.new_app('jenkins-ephemeral', current_project)
-    if(res == None):
-        print("Error while installing jenkins using ephemeral template")
-        raise AssertionError
-
 @then(u'route.route.openshift.io "jenkins" created')
 def checkRoute(context):
+    current_project = context.current_project
     try:
         res = oc.get_route('jenkins', current_project)
         if not 'jenkins' in res:
@@ -97,9 +78,9 @@ def checkRoute(context):
 Pre 4.6 configmap not available'
 '''
 
-
 @then(u'configmap "jenkins-trusted-ca-bundle" created')
 def checkConfigmap(context):
+    current_project = context.current_project
     try:
         res = oc.get_configmap(current_project)
         if not 'jenkins' in res:
@@ -108,21 +89,9 @@ def checkConfigmap(context):
     except AssertionError:
         print('Problem with configmap')
 
-
-@then(u'deploymentconfig.apps.openshift.io "jenkins" created')
-def checkDC(context):
-    try:
-        res = oc.get_deploymentconfig(current_project)
-        if not 'jenkins' in res:
-            raise AssertionError("deploymentconfig creation failed")
-        item = oc.search_resource_in_namespace('dc', 'jenkins', current_project)
-        print(f'deploymentconfig {item} created')
-    except AssertionError:
-        print('Problem with deploymentconfig')
-
-
 @then(u'serviceaccount "jenkins" created')
 def checkSA(context):
+    current_project = context.current_project
     try:
         res = oc.get_service_account(current_project)
         if not 'jenkins' in res:
@@ -135,6 +104,7 @@ def checkSA(context):
 
 @then(u'rolebinding.authorization.openshift.io "jenkins_edit" created')
 def checkRolebinding(context):
+    current_project = context.current_project
     try:
         res = oc.get_role_binding(current_project)
         if not 'jenkins' in res:
@@ -147,6 +117,7 @@ def checkRolebinding(context):
 
 @then(u'service "jenkins-jnlp" created')
 def checkSVCJNLP(context):
+    current_project = context.current_project
     try:
         res = oc.get_service(current_project)
         if not 'jenkins-jnlp' in res:
@@ -159,6 +130,7 @@ def checkSVCJNLP(context):
 
 @then(u'service "jenkins" created')
 def checkSVC(context):
+    current_project = context.current_project
     try:
         res = oc.get_service(current_project)
         if not 'jenkins' in res:
@@ -168,389 +140,17 @@ def checkSVC(context):
     except AssertionError:
         print(f'Problem with service jenkins')
 
-@then(u'We check for deployment pod status to be "Completed"')
-def deploymentPodStatus(context):
-    verify_status('pod', deploy_pod, poll_interval_seconds=2, max_retries=60, expected_status='Succeeded')
-
-
 @then(u'We check for jenkins master pod status to be "Ready"')
 def jenkinsMasterPodStatus(context):
-    global jenkins_master_pod
-    jenkins_master_pod = ""
-    jenkins_master_pod = getmasterpod(current_project)
+    current_project = context.current_project
+    jenkins_master_pod = oc.getmasterpod(current_project)
     print('---------Getting default jenkins pod name---------')
     print(jenkins_master_pod)
     container_status = oc.get_resource_info_by_jsonpath('pods',jenkins_master_pod,current_project,json_path='{.status.containerStatuses[*].ready}')
+    print(container_status)
     if 'false' in container_status:
         raise AssertionError
-    else:
-         print(container_status)
-
-@then(u'persistentvolumeclaim "jenkins" created')
-def verify_pvc(context):
-    if not 'jenkins' in oc.search_resource_in_namespace('pvc','jenkins',current_project):
-        raise AssertionError
-    else:
-        res = oc.search_resource_in_namespace('pvc','jenkins',current_project)
-        print(res)
-
-
-@then(u'we check the pvc status is "Bound"')
-def pvc_status(context):
-    print('---------Getting pvc status---------')
-    verify_status('pvc', 'jenkins', poll_interval_seconds=1, max_retries=10, expected_status='Bound')
-
 
 @given(u'The jenkins pod is up and runnning')
 def checkJenkins(context):
     jenkinsMasterPodStatus(context)
-
-@when(u'User enters oc new-app jenkins-persistent command')
-def persistentTemplate(context):
-    res = oc.new_app('jenkins-persistent', current_project)
-    if(res == None):
-        print("Error while installing jenkins using persistent template")
-        raise AssertionError
-
-@when(u'The user enters new-app command with nodejs_template')
-def createPipeline(context):
-    res = oc.new_app_from_file(nodejs_template,current_project)
-    if(res == None):
-        print("Error while installing nodejs using persistent nodejs_template")
-        raise AssertionError
-    time.sleep(30)
-    if 'sample-pipeline' in oc.search_resource_in_namespace('bc','sample', current_project):
-        print('Buildconfig sample-pipeline created')
-    elif 'nodejs-postgresql-example' in oc.search_resource_in_namespace('bc','postgersql',current_project):
-        print('Buildconfig nodejs-postgresql-example created')
-    else:
-        raise AssertionError
-    print(res)
-
-@then(u'Trigger the build using oc start-build')
-def startbuild(context):
-    triggerbuild('sample-pipeline',current_project)
-
-
-@then(u'verify the build status of "nodejs-postgresql-example-1" build is Complete')
-def verifynodejsBuildStatus(context):
-    verify_status('build', 'nodejs-postgresql-example-1', 5, 60, 'Complete')
-
-
-@then(u'verify the build status of "nodejs-postgresql-example-2" build is Complete')
-def verifynodejsBuildBStatus(context):
-    # give a little bit of time for the build to be created
-    time.sleep(60)
-    verify_status('build', 'nodejs-postgresql-example-2', 5, 60, 'Complete')
-
-
-@then(u'route nodejs-postgresql-example must be created and be accessible')
-def connectApp(context):
-    print('Getting application route/url')
-    app_name = 'nodejs-postgresql-example'
-    verify_status('route', 'nodejs-postgresql-example', 2, 10, 'True', json_path='{.status.ingress[*].conditions[*].status}')
-    route = oc.get_route_host(app_name,current_project)
-    url = 'http://'+str(route)
-    print('--->App url:')
-    print(url)
-    http = urllib3.PoolManager()
-    res = http.request_encode_url('GET',url)
-    connection_status = res.status
-    count = 1
-    while(count <= 30):
-        res = http.request_encode_url('GET',url)
-        connection_status = res.status
-        if connection_status == 200:
-            print('---> Application is accessible via the route')
-            print(url)
-            http.clear()
-            break
-        else:
-            time.sleep(2)
-            count+=1
-            print("Url: {0}, return code: {1}, res: {2}".format(url,connection_status, res))
-    if connection_status != 200:
-        raise AssertionError
-
-@when(u'The user create objects from the sample maven template by processing the template and piping the output to oc create')
-def createMavenTemplate(context):
-    res = oc.oc_process_template(maven_template)
-    print(res)
-
-@when(u'verify imagestream.image.openshift.io/openshift-jee-sample & imagestream.image.openshift.io/wildfly exist')
-def verifyImageStream(context):
-    if not 'openshift-jee-sample' in oc.search_resource_in_namespace('imagestream','openshift-jee-sample', current_project):
-        raise AssertionError
-    elif not 'wildfly' in oc.search_resource_in_namespace('imagestream','wildfly', current_project):
-        raise AssertionError
-    else:
-        res = oc.get_resource_lst('imagestream',current_project)
-        print(res)
-
-@when(u'verify buildconfig.build.openshift.io/openshift-jee-sample & buildconfig.build.openshift.io/openshift-jee-sample-docker exist')
-def verifyBuildConfig(context):
-    if not 'openshift-jee-sample' in oc.search_resource_in_namespace('buildconfig','openshift-jee-sample', current_project):
-        raise AssertionError
-    elif not 'openshift-jee-sample-docker' in oc.search_resource_in_namespace('buildconfig','openshift-jee-sample-docker', current_project):
-        raise AssertionError
-    else:
-        res = oc.get_resource_lst('buildconfig',current_project)
-        print(res)
-
-@when(u'verify deploymentconfig.apps.openshift.io/openshift-jee-sample is created')
-def verifyDeploymentConfig(context):
-    if not 'openshift-jee-sample' in oc.search_resource_in_namespace('deploymentconfig','openshift-jee-sample',current_project):
-        raise AssertionError
-    else:
-        res = oc.search_resource_in_namespace('deploymentconfig','openshift-jee-sample',current_project)
-        print(res)
-
-@when(u'verify service/openshift-jee-sample is created')
-def verifySvc(context):
-    if not 'openshift-jee-sample' in oc.search_resource_in_namespace('service','openshift-jee-sample',current_project):
-        raise AssertionError
-    else:
-        res = oc.search_resource_in_namespace('service','openshift-jee-sample',current_project)
-        print(res)
-
-@when(u'verify route.route.openshift.io/openshift-jee-sample is created')
-def verifyRoute(context):
-    if not 'openshift-jee-sample' in oc.search_resource_in_namespace('route','openshift-jee-sample',current_project):
-        raise AssertionError
-    else:
-        res = oc.search_resource_in_namespace('route','openshift-jee-sample',current_project)
-        print(res)
-
-@then(u'Trigger the build using oc start-build openshift-jee-sample')
-def startBuild(context):
-    triggerbuild('openshift-jee-sample',current_project)
-
-@then(u'verify the build status of openshift-jee-sample-docker build is Complete')
-def verifyDockerBuildStatus(context):
-    verify_status('build', 'openshift-jee-sample-docker-1', 2, 10, 'Complete', '{.status.phase}')
-
-
-@then(u'verify the build status of openshift-jee-sample-1 is Complete')
-def verifyJenkinsBuildStatus(context):
-    verify_status('build', 'openshift-jee-sample-1', 2, 100, 'Complete', '{.status.phase}')
-
-
-def verify_status(object_type, object_name, poll_interval_seconds, max_retries, expected_status,json_path='{.status.phase}'):
-    count = 1
-    print("Getting {object_type} status for {object_name}")
-    while(count <= max_retries):
-        status = oc.get_resource_info_by_jsonpath(object_type,object_name,current_project,json_path)
-        if expected_status in status:
-            break
-        time.sleep(poll_interval_seconds)
-        count+=1
-    print("{object_type} {object_name} status:{status}")
-    if not expected_status in status:
-        raise AssertionError
-
-@then(u'verify the JaveEE application is accessible via route openshift-jee-sample')
-def pingApp(context):
-    print('Getting application route/url')
-    app_name = 'openshift-jee-sample'
-    route = oc.get_route_host(app_name,current_project)
-    url = 'http://'+str(route)
-    print('--->App url:')
-    print(url)
-    http = urllib3.PoolManager()
-    res = http.request_encode_url('GET',url)
-    connection_status = res.status
-    if connection_status == 200:
-        print('---> Application is accessible via the route')
-        print(url)
-        http.clear()
-    else:
-        raise Exception
-
-@then(u'we delete deploymentconfig.apps.openshift.io "jenkins"')
-def del_dc(context):
-    global jenkins_master_pod
-    jenkins_master_pod = ''
-    res = oc.delete("deploymentconfig","jenkins",current_project)
-    if res == None:
-        raise AssertionError
-
-@then(u'we delete route.route.openshift.io "jenkins"')
-def del_route(context):
-    res = oc.delete("route","jenkins",current_project)
-    if res == None:
-        raise AssertionError
-
-
-@then(u'delete configmap "jenkins-trusted-ca-bundle"')
-def del_cm(context):
-    res = oc.delete("configmap","jenkins-trusted-ca-bundle",current_project)
-    if res == None:
-        raise AssertionError
-
-
-@then(u'delete serviceaccount "jenkins"')
-def del_sa(context):
-    res = oc.delete("serviceaccount","jenkins",current_project)
-    if res == None:
-        raise AssertionError
-
-
-@then(u'delete rolebinding.authorization.openshift.io "jenkins_edit"')
-def del_rb(context):
-    res = oc.delete("rolebinding","jenkins_edit",current_project)
-    if res == None:
-        raise AssertionError
-
-@then(u'delete service "jenkins"')
-def del_svc(context):
-    res = oc.delete("service","jenkins",current_project)
-    if res == None:
-        raise AssertionError
-
-@then(u'delete service "jenkins-jnlp"')
-def del_svc_jnlp(context):
-    res = oc.delete("service","jenkins-jnlp",current_project)
-    if res == None:
-        raise AssertionError
-
-@then(u'delete all buildconfigs')
-def del_bc(context):
-    res = oc.delete("bc","--all",current_project)
-    if res == None:
-        raise AssertionError
-
-@then(u'delete all builds')
-def del_builds(context):
-    res = oc.delete("builds","--all",current_project)
-    if res == None:
-        raise AssertionError
-
-@then(u'delete all deploymentconfig')
-def del_alldc(context):
-    res = oc.delete("deploymentconfig","--all",current_project)
-    if res == None:
-        raise AssertionError
-
-@then(u'delete all build pods')
-def del_pods(context):
-    pods = v1.list_namespaced_pod(current_project)
-    buildpods = []
-    for i in pods.items:
-        if 'jenkins-1-deploy' not in i.metadata.name and jenkins_master_pod not in i.metadata.name:
-            buildpods.append(i.metadata.name)
-    for pod in buildpods:
-        res = oc.delete('pod',pod,current_project)
-        print("Deleting: ",res)
-
-@then(u'delete all remaining test resources')
-def del_all_remaining_test_resources(context):
-    #    oc delete all,rolebindings.authorization.openshift.io,bc,cm,is,pvc,sa,secret -l app=jenkins-ephemeral
-    #    oc delete all,rolebindings.authorization.openshift.io,bc,cm,is,pvc,sa,secret -l app=jenkins-persistent
-    #    oc delete all,rolebindings.authorization.openshift.io,bc,cm,is,pvc,sa,secret -l app=openshift-jee-sample
-    #    oc delete all,rolebindings.authorization.openshift.io,bc,cm,is,pvc,sa,secret -l app=jenkins-pipeline-example
-    delete_command = "all,rolebindings.authorization.openshift.io,bc,cm,is,pvc,sa,secret"
-    oc.delete(delete_command,"-l app=jenkins-ephemeral",current_project)
-    oc.delete(delete_command,"-l app=jenkins-persistent",current_project)
-    oc.delete(delete_command,"-l app=openshift-jee-sample",current_project)
-    oc.delete(delete_command,"-l app=jenkins-pipeline-example",current_project)
-
-@then(u'We rsh into the master pod and check the jobs count')
-def getjobcount(context):
-    for jobnames,_ in buildconfigs.items():
-        exec_command = 'cat /var/lib/jenkins/jobs/'+current_project+'/jobs/'+current_project+'-'+jobnames+'/nextBuildNumber'
-        count = oc.exec_in_pod(jenkins_master_pod,exec_command)
-        buildconfigs[jobnames] = str(count)
-    print(buildconfigs)
-
-@when(u'We delete the jenkins master pod')
-def deletemaster(context):
-    master_pod = getmasterpod(current_project)
-    res = oc.delete("pods",master_pod,current_project)
-    time.sleep(60)
-    if res == None:
-        raise AssertionError
-
-@then(u'We rsh into the master pod & Compare if the data persist or is lost upon pod restart')
-def comparejobs(context):
-    for jobnames,_ in buildconfigs.items():
-        master_pod = getmasterpod(current_project)
-        exec_command = 'cat /var/lib/jenkins/jobs/'+current_project+'/jobs/'+current_project+'-'+jobnames+'/nextBuildNumber'
-        count = oc.exec_in_pod(master_pod,exec_command)
-        buildconfigs[jobnames] = str(count)
-    
-    for jobnames, _ in buildconfigs.items():
-        if(buildconfigs[jobnames] == '1'):
-            print("Data doesnt persist")
-            raise AssertionError
-    print(buildconfigs)
-
-
-def getmasterpod(namespace: str)-> str:
-    '''
-    returns the jenkins master pod name
-    '''
-    pods = v1.list_namespaced_pod(current_project)
-    for i in pods.items:
-        if 'jenkins-1-deploy' not in i.metadata.name and 'jenkins-1' in i.metadata.name:
-            master_pod = i.metadata.name
-    return str(master_pod)
-
-@when(u'We rsh into the master pod')
-def step_impl(context):
-    pass
-
-@then(u'We compare the plugins version inside the master pod with the plugins listed in plugins.txt')
-def step_impl(context):
-    pass
-
-@when(u'We Trigger multiple builds using oc start-build openshift-jee-sample')
-def multiplebuilds(context):
-    global builds
-    count = 1
-    ## creating a dictionary of builds that keeps a track of {buildname: build_status}
-         # This will be used to check the build reconcilation
-    while(count <= 5):
-        triggerbuild('openshift-jee-sample',current_project)
-        build_name ='openshift-jee-sample-' + str(count)
-        builds[build_name] = None
-        count+=1
-    
-@when(u'We scale down the pod count in the replication controller to "0" from "1"')
-def podscaling(context):
-    rc_name = 'jenkins-1'
-    oc.scaleReplicas(current_project,0,rc_name)
-    replicas = oc.get_resource_info_by_jsonpath("dc","jenkins",current_project,json_path='{.status.availableReplicas}')
-    if not '0' in replicas:
-        raise AssertionError
-    else:
-        print('There are ',replicas,' running pods of jenkins')
-    
-@then(u'We delete some builds')
-def deletebuilds(context):
-    global builds
-    rm_build = ['openshift-jee-sample-2','openshift-jee-sample-4']
-    for build_name in builds.keys():
-        builds[build_name] = oc.get_resource_info_by_jsonpath("build",build_name,current_project,json_path='{.status.phase}')
-    print("------------Fetching all builds and build status------------")
-    print(builds)
-    print("------------Deleting a few  builds------------")
-    for items in rm_build:
-        res = oc.delete("build",items,current_project)
-        print(res)
-        builds.pop(items)
-    print("------------Fetching all builds and build status------------")
-    print(builds)
-    # This sleep is used to give time for the jenkins master pod to be back
-    time.sleep(60)
-
-
-@then(u'verify sync plugin is able to reconcile the build state and delete the job runs associated with the builds we deleted')
-def verifybuildSync(context):
-    time.sleep(360)
-    for build_name in builds.keys():
-        builds[build_name] = oc.get_resource_info_by_jsonpath("build",build_name,current_project,json_path='{.status.phase}')
-        if not "Complete" or "New" in builds[build_name]:
-            print(build_name,':',builds[build_name])
-            raise AssertionError
-        else:
-            print(build_name,':',builds[build_name])
