+++ smoke/features/stress-test.feature
@@ -8,5 +8,5 @@ Feature: To test sync plugin performance and stability on different load
       When We Trigger multiple builds using oc start-build openshift-jee-sample
       And We scale down the pod count in the replication controller to "0" from "1"
       Then We delete some builds
-      Then We check for jenkins master pod status to be "Ready"
+      Then We ensure that jenkins deployment config is ready
       And verify sync plugin is able to reconcile the build state and delete the job runs associated with the builds we deleted
