+++ smoke/requirements.txt
@@ -1,8 +1,9 @@
-
 behave
 pyshould
 requests
-kubernetes == 12.0.1
+kubernetes
 urllib3 == 1.25.9
 pyfiglet
 clint
+python-jenkins
+
